refactor: restructure shared-contrls modules
