+++ superset-frontend/packages/superset-ui-chart-controls/src/utils/expandControlConfig.tsx
@@ -17,8 +17,7 @@
  * under the License.
  */
 import React, { ReactElement } from 'react';
-import sharedControls from '../shared-controls';
-import sharedControlComponents from '../shared-controls/components';
+import { sharedControls, sharedControlComponents } from '../shared-controls';
 import {
   ControlType,
   ControlSetItem,
